Encapsulate the search city functionality separately.
