+++ src/components/dashboard/MapContext.tsx
@@ -1,0 +1,5 @@
+// MapContext.tsx
+import React from "react";
+import mapboxgl from "mapbox-gl";
+
+export const MapContext = React.createContext<mapboxgl.Map | null>(null);
